change total_mark column data type to tiny int to interger

+++ database/migrations/2025_08_12_193402_create_recognitions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
 
             $table->foreignId('user_id')->constrained()->onDelete('cascade')->onUpdate('cascade');
             $table->string('category');
-            $table->mediumInteger('total_mark');
+            $table->integer('total_mark');
             $table->text('comment');
             $table->foreignId('recognizer_id')->constrained('users')->onDelete('cascade');
